refactor: extract header component
